[FIX] Function Flow: Coin Selection Mismatch and Proposed Fixes #2810 (#2813)
* Enhance FunctionCallViewModel and FunctionCallDetailsView to ensure RUNE token is selected for THORChain operations. Added setRuneToken method to handle token selection for BOND, UNBOND, LEAVE, MERGE, ADD THORCHAIN LP, and REMOVE THORCHAIN LP functions.

* Refactor FunctionCall classes to include vault parameter and ensure RUNE token selection for BOND, UNBOND, and LEAVE operations on THORChain. Updated FunctionCallDetailsView to pass vault to instances.

* Add isInThePool flag to FunctionCallAddThorLP for improved RUNE token selection logic

This update introduces a new boolean variable, isInThePool, to track if the selected coin is part of the available pools. The logic ensures that the RUNE token is set correctly when the selected coin is not RUNE and is not found in the pool options. This enhancement improves the handling of THORChain operations in the FunctionCallAddThorLP class.

---------

Co-authored-by: hailhydra <johnnyluo1980@gmail.com>